move lines command into own package
