+++ src/main/java/com/fossgalaxy/pircbotx/commands/lines/LineModel.java
@@ -16,7 +16,7 @@
  * You should have received a copy of the GNU General Public License along with
  * uc_pircbotx. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.fossgalaxy.pircbotx.data.db;
+package com.fossgalaxy.pircbotx.commands.lines;
 
 import com.google.inject.Inject;
 import org.slf4j.Logger;
