+++ src/main/java/com/fossgalaxy/pircbotx/commands/lines/LinesListener.java
@@ -16,9 +16,8 @@
  * You should have received a copy of the GNU General Public License along with
  * uc_pircbotx. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.fossgalaxy.pircbotx.listeners;
+package com.fossgalaxy.pircbotx.commands.lines;
 
-import com.fossgalaxy.pircbotx.data.db.LineModel;
 import com.google.inject.Inject;
 import org.pircbotx.hooks.ListenerAdapter;
 import org.pircbotx.hooks.events.MessageEvent;
