+++ src/main/java/com/fossgalaxy/pircbotx/commands/lines/RandCommand.java
@@ -16,11 +16,10 @@
  * You should have received a copy of the GNU General Public License along with
  * uc_pircbotx. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.fossgalaxy.pircbotx.commands;
+package com.fossgalaxy.pircbotx.commands.lines;
 
 import com.fossgalaxy.pircbotx.commandprocessor.Command;
 import com.fossgalaxy.pircbotx.commandprocessor.Message;
-import com.fossgalaxy.pircbotx.data.db.LineModel;
 import com.fossgalaxy.pircbotx.modules.AnnotationModule;
 import com.google.inject.Inject;
 import org.slf4j.Logger;
