+++ src/main/resources/META-INF/services/com.fossgalaxy.pircbotx.modules.Module
@@ -1,5 +1,5 @@
 com.fossgalaxy.pircbotx.commands.CalcCommand
-com.fossgalaxy.pircbotx.commands.RandCommand
+com.fossgalaxy.pircbotx.commands.lines.RandCommand
 com.fossgalaxy.pircbotx.commands.DateTimeCommand
 com.fossgalaxy.pircbotx.commands.KillerTroutCommand
 com.fossgalaxy.pircbotx.commands.joins.JoinsCommand
